+++ refinery-spark/src/main/scala/org/wikimedia/analytics/refinery/spark/sql/JsonSchemaConverter.scala
@@ -2,6 +2,7 @@ package org.wikimedia.analytics.refinery.spark.sql
 
 import com.fasterxml.jackson.databind.JsonNode
 import org.apache.spark.sql.types._
+import org.wikimedia.analytics.refinery.core.LogHelper
 
 import scala.collection.JavaConverters._
 import scala.util.Try
@@ -13,7 +14,7 @@ import scala.util.Try
  * - Only a subset of JSONSchema is supported.
  * - object with additionalProperties only are converted to a Map.
  */
-object JsonSchemaConverter {
+object JsonSchemaConverter extends LogHelper {
 
     // Useful JSONSchema field names
     private val typeField                    = "type"
@@ -35,10 +36,10 @@ object JsonSchemaConverter {
         val itemsSchema = jsonSchema.get(itemsField)
         // Arrays must specify the type of their elements.
         if (itemsSchema == null || itemsSchema.isNull)
-            throw new IllegalArgumentException(s"$fieldName array schema did not specify the items field")
+            throw new IllegalArgumentException(s"`$fieldName` array schema did not specify the items field")
         // Arrays must only use a single type, not tuple validation.
         if (!itemsSchema.isObject || !itemsSchema.has(typeField)) {
-            throw new IllegalArgumentException(s"$fieldName array schema must specify the items type field")
+            throw new IllegalArgumentException(s"`$fieldName` array schema must specify the items type field")
         }
         ArrayType(buildDataType(itemsSchema, s"$fieldName-$propertiesField"), containsNull = false)
     }
@@ -54,10 +55,10 @@ object JsonSchemaConverter {
         val additionalPropertiesSchema = jsonSchema.get(additionalPropertiesField)
         if (additionalPropertiesSchema == null || additionalPropertiesSchema.isNull)
             throw new IllegalArgumentException(
-                s"$fieldName object schema must specify either properties or additionalProperties field")
+                s"`$fieldName` object schema must specify either properties or additionalProperties field")
         if (!additionalPropertiesSchema.isObject || !additionalPropertiesSchema.has(typeField))
             throw new IllegalArgumentException(
-                s"$fieldName object schema additionalProperties field must specify its type")
+                s"`$fieldName` object schema additionalProperties field must specify its type")
         MapType(
             StringType,
             buildDataType(additionalPropertiesSchema, s"$fieldName-$additionalPropertiesField"),
@@ -75,19 +76,19 @@ object JsonSchemaConverter {
      */
     private def buildDataType(jsonSchema: JsonNode, fieldName: String): DataType = {
         if (jsonSchema == null ||  jsonSchema.isNull)
-            throw new IllegalArgumentException("Schema node shouldn't be null when building DataType")
+            throw new IllegalArgumentException(s"$fieldName` node shouldn't be null when building DataType")
         if (!jsonSchema.isObject)
-            throw new IllegalArgumentException("Schema node should be an object when building DataType")
+            throw new IllegalArgumentException(s"$fieldName` node should be an object when building DataType")
 
         val schemaTypeNode = jsonSchema.get(typeField)
         if (schemaTypeNode == null || !schemaTypeNode.isTextual)
-            throw new IllegalArgumentException("Schema must contain 'type' field")
+            throw new IllegalArgumentException(s"`$fieldName` schema must contain 'type' field")
 
         val typeText = schemaTypeNode.textValue()
         typeText match {
             case "null" => NullType
             case "boolean" => BooleanType
-            case "integer" => IntegerType
+            case "integer" => LongType
             case "number" => DoubleType
             case "string" => StringType
             case "array" => buildArrayType(jsonSchema, fieldName)
@@ -97,7 +98,7 @@ object JsonSchemaConverter {
                 if (properties != null && !properties.isNull) buildStructType(jsonSchema, fieldName)
                 // Special map-case: properties undefined, additionalProperties defined with type
                 else buildMapType(jsonSchema, fieldName)
-            case _ => throw new IllegalArgumentException(s"$fieldName has invalid type value: $typeText")
+            case _ => throw new IllegalArgumentException(s"`$fieldName` has invalid type value: $typeText")
         }
     }
 
@@ -109,7 +110,13 @@ object JsonSchemaConverter {
      */
     private def buildStructField(jsonSchema: JsonNode, fieldName: String, required: Boolean): StructField = {
         val datatype = buildDataType(jsonSchema, fieldName)
-        val field = StructField(fieldName, datatype, nullable = !required)
+        log.debug(s"Converting JSONSchema field `$fieldName` to Spark dataType $datatype")
+
+        // If this field is in the parent's required list
+        // or it is marked as a draft 3 required field,
+        // then it should not be nullable.
+        val nullable = !(required || isRequiredDraft3(jsonSchema))
+        val field = StructField(fieldName, datatype, nullable = nullable)
 
         val description = jsonSchema.get(descriptionField)
         // Return field with comment if defined, without otherwise
@@ -128,17 +135,37 @@ object JsonSchemaConverter {
     private def buildStructType(jsonSchema: JsonNode, fieldName: String = "root"): StructType = {
         val properties = jsonSchema.get(propertiesField)
         if (properties == null || !properties.isObject)
-            throw new IllegalArgumentException(fieldName + " struct schema's \"properties\" is not an object.")
-        val requiredFieldList = jsonSchema.get(requiredField)
-        if (requiredFieldList != null && !requiredFieldList.isArray)
-            throw new IllegalArgumentException(fieldName + " struct schema's \"required\" is not an array.")
-        val requiredFieldSet = Try(requiredFieldList.elements().asScala.map(_.asText()).toSet).getOrElse(Set.empty)
+            throw new IllegalArgumentException(s"`$fieldName` struct schema's properties is not an object.")
+
+        val requiredFieldValue = jsonSchema.get(requiredField)
+        // If required is set, it must either be an array (Draft 4+) or a boolean (Draft 3)
+        if (requiredFieldValue != null && !(requiredFieldValue.isArray || requiredFieldValue.isBoolean))
+            throw new IllegalArgumentException(
+                s"`$fieldName` struct schema's required is not an array or boolean (JSONSchema Draft3)."
+            )
+
+        val requiredFieldSet = Try(requiredFieldValue.elements().asScala.map(_.asText()).toSet).getOrElse(Set.empty)
         val fields = properties.fields.asScala.toSeq
+
         StructType(fields.map(mapEntry => {
             val subfieldName = mapEntry.getKey
             val subfieldNode = mapEntry.getValue
             buildStructField(subfieldNode, subfieldName, requiredFieldSet.contains(subfieldName))
         }))
+    }
+
+    /**
+      * If a field has a 'required' property as a boolean, assume this is a JSONSchema Draft3
+      * style schema and the required property is refering to the current field, not a list of
+      * required sub properties.  This returns true if the jsonSchema has required: true
+      *
+      * @param jsonSchema
+      * @return
+      */
+    private def isRequiredDraft3(jsonSchema: JsonNode): Boolean = {
+        jsonSchema.has("required") &&
+            jsonSchema.get("required").isBoolean &&
+            jsonSchema.get("required").booleanValue
     }
 
     /**